dom/media/tests/mochitest/test_getUserMedia_constraints.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=882145
-->
<head>
  <meta charset="utf-8">
  <title>Test mozGetUserMedia Constraints</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=882145">Test mozGetUserMedia Constraints</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">
/**
  Tests covering gUM constraints API for audio, video and fake video. Exercise
  successful parsing code and ensure that unknown required constraints and
  overconstraining cases produce appropriate errors.
*/
var tests = [
  // Each test here tests a different constraint or codepath.
  { message: "unknown required constraint on video fails",
    constraints: { video: { somethingUnknown:0, require:["somethingUnknown"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "unknown required constraint on audio fails",
    constraints: { audio: { somethingUnknown:0, require:["somethingUnknown"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "missing required constraint on video fails",
    constraints: { video: { require:["facingMode"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "missing required constraint on audio fails",
    constraints: { audio: { require:["facingMode"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "video overconstrained by facingMode fails",
    constraints: { video: { facingMode:'left', require:["facingMode"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "audio overconstrained by facingMode fails",
    constraints: { audio: { facingMode:'left', require:["facingMode"] } },
    error: "NO_DEVICES_FOUND",
    pass: false },
  { message: "Success-path: optional video facingMode + audio ignoring facingMode",
    constraints: { fake: true,
                   audio: { facingMode:'left',
                            foo:0,
                            advanced: [{ facingMode:'environment' },
                                       { facingMode:'user' },
                                       { bar:0 }] },
                   video: { // TODO: Bug 767924 sequences in unions
                            //facingMode:['left', 'right', 'user', 'environment'],
                            //require:["facingMode"],
                            facingMode:'left',
                            foo:0,
                            advanced: [{ facingMode:'environment' },
                                       { facingMode:'user' },
                                       { bar:0 }] } },
    error: null,
    pass: false },
  { message: null },
];

/**
 * Starts the test run by running through each constraint
 * test by verifying that the right callback and error message is fired.
 */

runTest(function () {
  var i = 0;
  next();

  function Success() {
    info("successcallback");
    tests[i].pass = !tests[i].error;
    i++;
    next();
  }
  function Failure(err) {
    info("errcallback: " + err);
    tests[i].pass = tests[i].error? (err === tests[i].error) : false;
    i++;
    next();
  }
  function next() {
    if (tests[i].message) {
      navigator.mozGetUserMedia(tests[i].constraints, Success, Failure);
    } else {
      finish();
    }
  }
  function finish() {
    tests.forEach(function (test) {
      if (test.message) {
        ok(test.pass, test.message);
      } else {
        SimpleTest.finish();
      }
    });
  }
});


</script>
</pre>
</body>
</html>
